requirements.txt
Flask==3.0.3
feedparser==6.0.11
newspaper3k==0.2.8
nltk==3.8.1
requests==2.27.1
python-dotenv==0.20.0
numpy==1.21.5
sumy==0.11.0
